feat(profile-widget): add generic profile widget
This commit introduces a new generic profile widget component to be included in the site's layout. The component displays a user's avatar, name, optionally a prefix and postfix for the name, and an affiliation with or without a URL. It's designed to be flexible for various user profiles.

- A new HTML snippet `_includes/profile-widget-generic.html` has been created.
- SCSS partial `_sass/components/_profile_widget_generic.scss` has been added for styling the widget.
- The new SCSS file is imported into the main components stylesheet `_sass/_components.scss`.

The profile widget is structured with flexbox for a responsive layout and includes appropriate fallbacks for the images and links if certain data (like urls) are not provided. The SCSS uses variables and modifiers for easy theming and adjustment.